[Dropdown] Move new Dropdown to `navigation-dropdown` and revert old Dropdown
