+++ assets/javascripts/kitten/components/navigation/header-nav/components/burger-menu.js
@@ -1,6 +1,6 @@
 import React, { useRef } from 'react'
 import { BurgerIcon } from '../../../../components/icons/burger-icon'
-import { Dropdown } from '../../../../components/dropdowns/dropdown'
+import { Dropdown } from '../../../../components/dropdowns/navigation-dropdown/dropdown'
 import { VisuallyHidden } from '../../../../components/accessibility/visually-hidden'
 import COLORS from '../../../../constants/colors-config'
 import { Context } from './context'
